+++ apps/dialog/src/app/api/chat/chat-service.ts
@@ -32,7 +32,7 @@ import { extractUrls } from '../utils/extract-urls';
 import { UserAndContext } from '@/auth/types';
 import { extractImagesAndUrl } from '../file-operations/preprocess-image';
 import { ingestWebContent } from '../rag/ingestWebContent';
-import { searchWeb } from './websearch-service';
+import { searchWeb } from './websearch';
 
 /**
  * Converts frontend messages to ai-core message format
